pyproject.toml
[project]
name = "helix"
version = "0.1.0"
description = "Enterprise AI Agent Orchestration Platform"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Core
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14",
    "redis[hiredis]>=5.2",
    "celery[redis]>=5.4",
    "httpx>=0.28",
    # AI
    "langgraph>=0.3",
    "pydantic-ai>=0.0.36",
    "anthropic>=0.42",
    "openai>=1.58",
    # Auth
    "workos>=5.0",
    "python-jose[cryptography]>=3.3",
    # Integrations
    "composio-core>=0.7",
    # Observability
    "structlog>=24.4",
    "sentry-sdk[fastapi]>=2.19",
    "opentelemetry-api>=1.29",
    # Utils
    "python-dotenv>=1.0",
    "python-multipart>=0.0.18",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
    "mypy>=1.13",
    "factory-boy>=3.3",
    "testcontainers[postgres,redis]>=4.8",
    "httpx",  # for FastAPI TestClient
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/helix"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "T20"]
ignore = ["E501", "B008"]  # B008: Depends() in defaults is standard FastAPI pattern

[tool.ruff.lint.isort]
known-first-party = ["helix"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
filterwarnings = ["ignore::DeprecationWarning"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
